requirements.txt
altair==5.5.0
anyio==4.10.0
asttokens==3.0.0
attrs==21.4.0
blinker==1.9.0
cachetools==6.2.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
commonmark==0.9.1
decorator==5.2.1
et_xmlfile==2.0.0
executing==2.2.1
fastapi==0.117.1
fire==0.4.0
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
idna==3.10
importlib-metadata==3.3.0
interchange==2021.0.4
ipython==9.5.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jsonpickle==4.1.1
jsonschema==4.17.3
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
monotonic==1.6
narwhals==2.5.0
neo4j==5.28.2
networkx==3.5
numpy==2.3.3
openpyxl==3.1.5
packaging==25.0
pandas==2.3.2
pansi==2024.11.0
parso==0.8.5
pexpect==4.9.0
pillow==11.3.0
plotly==6.3.0
prompt_toolkit==3.0.52
protobuf==6.32.1
ptyprocess==0.7.0
pure_eval==0.2.3
py2neo==2021.2.4
pyarrow==21.0.0
pyattck==7.1.2
pyattck-data==2.6.3
pydantic==1.10.23
pydeck==0.9.1
Pygments==2.19.2
pyrsistent==0.20.0
python-dateutil==2.9.0.post0
pytz==2025.2
pyvis==0.3.2
PyYAML==6.0.2
requests==2.32.5
rich==12.6.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
stack-data==0.6.3
starlette==0.48.0
streamlit==1.49.1
tenacity==9.1.2
termcolor==3.1.0
toml==0.10.2
tornado==6.5.2
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.36.0
wcwidth==0.2.13
zipp==3.23.0
